fix(Interaction): ensure haptic pulse duration is a float
Previously, the haptic pulse duration on the interact touch/grab/use
scripts was being cast to an int, which is incorrect, the type taken
byt the `TriggerHapticPulse` method is of type float so no cast
is needed.

+++ Assets/SteamVR_Unity_Toolkit/Scripts/VRTK_InteractGrab.cs
@@ -401,7 +401,7 @@ namespace VRTK
                     var rumbleAmount = grabbedObject.GetComponent<VRTK_InteractableObject>().rumbleOnGrab;
                     if (!rumbleAmount.Equals(Vector2.zero))
                     {
-                        controllerActions.TriggerHapticPulse((ushort)rumbleAmount.y, (int)rumbleAmount.x, 0.05f);
+                        controllerActions.TriggerHapticPulse((ushort)rumbleAmount.y, rumbleAmount.x, 0.05f);
                     }
                 }
             }
